refactor: move the serializer beside the rest of the package
It sits with null.py, route.py and the other modules of the SDK rather than
under utils, which holds the two helpers that back generated resources.
